+++ tasks/workflow/worker.go
@@ -128,7 +128,7 @@ func getOperationExecutor(kv *api.KV, deploymentID, artifact string) (prov.Opera
 	return nil, originalErr
 }
 
-func (w *worker) monitorTaskCancellation(ctx context.Context, cancelFunc context.CancelFunc, t *taskExecution) {
+func (w *worker) monitorTaskCancellation(ctx context.Context, cancelFunc context.CancelFunc, t *taskExecution, workflowName string) {
 	go func() {
 		var lastIndex uint64
 		for {
@@ -149,6 +149,7 @@ func (w *worker) monitorTaskCancellation(ctx context.Context, cancelFunc context
 				if strings.ToLower(string(kvp.Value)) == "true" {
 					log.Debugln("Task cancellation requested.")
 					checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusCANCELED)
+					events.PublishAndLogWorkflowStatusChange(ctx, w.consulClient.KV(), t.targetID, t.taskID, workflowName, tasks.TaskStatusCANCELED.String())
 					cancelFunc()
 					return
 				}
@@ -284,7 +285,7 @@ func (w *worker) handleExecution(t *taskExecution) {
 		log.Printf("%+v", err)
 		return
 	}
-	w.monitorTaskCancellation(ctx, cancelFunc, t)
+	w.monitorTaskCancellation(ctx, cancelFunc, t, wfName)
 	w.monitorTaskFailure(ctx, cancelFunc, t)
 	defer func(t *taskExecution, start time.Time) {
 		if taskStatus, err := t.getTaskStatus(); err != nil && taskStatus != tasks.TaskStatusRUNNING {
@@ -298,7 +299,14 @@ func (w *worker) handleExecution(t *taskExecution) {
 	case tasks.TaskTypeUnDeploy, tasks.TaskTypePurge:
 		w.runUndeploy(ctx, t)
 	case tasks.TaskTypeCustomCommand:
-		w.runCustomCommand(ctx, t)
+		err := w.runCustomCommand(ctx, t)
+		if err != nil {
+			log.Printf("Deployment id: %q, Task id: %q, Failed to get Custom command name: %+v", t.targetID, t.taskID, err)
+			checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusFAILED)
+			events.PublishAndLogCustomCommandStatusChange(ctx, t.kv, t.targetID, t.taskID, tasks.TaskStatusFAILED.String())
+			return
+		}
+		events.PublishAndLogCustomCommandStatusChange(ctx, t.kv, t.targetID, t.taskID, tasks.TaskStatusDONE.String())
 	case tasks.TaskTypeScaleOut:
 		w.runScaleOut(ctx, t)
 	case tasks.TaskTypeScaleIn:
@@ -316,30 +324,22 @@ func (w *worker) handleExecution(t *taskExecution) {
 	}
 }
 
-func (w *worker) runCustomCommand(ctx context.Context, t *taskExecution) {
+func (w *worker) runCustomCommand(ctx context.Context, t *taskExecution) error {
 	kv := w.consulClient.KV()
 	commandName, err := tasks.GetTaskData(kv, t.taskID, "commandName")
 	if err != nil {
-		log.Printf("Deployment id: %q, Task id: %q, Failed to get Custom command name: %+v", t.targetID, t.taskID, err)
-		checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusFAILED)
-		return
+		return err
 	}
 	interfaceNameKv, _, err := kv.Get(path.Join(consulutil.TasksPrefix, t.taskID, "interfaceName"), nil)
 	if err != nil {
-		log.Printf("Deployment id: %q, Task id: %q, Failed to get Custom command name: %+v", t.targetID, t.taskID, err)
-		checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusFAILED)
-		return
+		return err
 	}
 	nodes, err := tasks.GetTaskRelatedNodes(kv, t.taskID)
 	if err != nil {
-		log.Printf("Deployment id: %q, Task id: %q, Failed to get Custom command node: %+v", t.targetID, t.taskID, err)
-		checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusFAILED)
-		return
+		return err
 	}
 	if len(nodes) != 1 {
-		log.Printf("Deployment id: %q, Task id: %q, Expecting custom command TaskExecution to be related to \"1\" node while it is actually related to \"%d\" nodes", t.targetID, t.taskID, len(nodes))
-		checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusFAILED)
-		return
+		return errors.Errorf("Deployment id: %q, Task id: %q, Expecting custom command TaskExecution to be related to \"1\" node while it is actually related to \"%d\" nodes", t.targetID, t.taskID, len(nodes))
 	}
 	nodeName := nodes[0]
 	interfaceName := "custom"
@@ -348,29 +348,23 @@ func (w *worker) runCustomCommand(ctx context.Context, t *taskExecution) {
 	}
 	nodeType, err := deployments.GetNodeType(w.consulClient.KV(), t.targetID, nodeName)
 	if err != nil {
-		log.Printf("Deployment id: %q, Task id: %q, Failed to get Custom command node type: %+v", t.targetID, t.taskID, err)
-		checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusFAILED)
-		return
+		return err
 	}
 	op, err := operations.GetOperation(ctx, kv, t.targetID, nodeName, interfaceName+"."+commandName, "", "")
 	if err != nil {
-		log.Printf("Deployment id: %q, Task id: %q, Command TaskExecution failed for node %q: %+v", t.targetID, t.taskID, nodeName, err)
 		err = setNodeStatus(ctx, t.kv, t.taskID, t.targetID, nodeName, tosca.NodeStateError.String())
 		if err != nil {
 			log.Printf("Deployment id: %q, Task id: %q, Failed to set status for node %q: %+v", t.targetID, t.taskID, nodeName, err)
 		}
-		checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusFAILED)
-		return
+		return errors.Wrapf(err, "Command TaskExecution failed for node %q", nodeName)
 	}
 	exec, err := getOperationExecutor(kv, t.targetID, op.ImplementationArtifact)
 	if err != nil {
-		log.Printf("Deployment id: %q, Task id: %q, Command TaskExecution failed for node %q: %+v", t.targetID, t.taskID, nodeName, err)
 		err = setNodeStatus(ctx, t.kv, t.taskID, t.targetID, nodeName, tosca.NodeStateError.String())
 		if err != nil {
 			log.Printf("Deployment id: %q, Task id: %q, Failed to set status for node %q: %+v", t.targetID, t.taskID, nodeName, err)
 		}
-		checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusFAILED)
-		return
+		return errors.Wrapf(err, "Command TaskExecution failed for node %q", nodeName)
 	}
 
 	ctx = operations.SetOperationLogFields(ctx, op)
@@ -382,16 +376,15 @@ func (w *worker) runCustomCommand(ctx context.Context, t *taskExecution) {
 	}()
 	if err != nil {
 		metrics.IncrCounter(metricsutil.CleanupMetricKey([]string{"executor", "operation", t.targetID, nodeType, op.Name, "failures"}), 1)
-		log.Printf("Deployment id: %q, Task id: %q, Command TaskExecution failed for node %q: %+v", t.targetID, t.taskID, nodeName, err)
 		err = setNodeStatus(ctx, t.kv, t.taskID, t.targetID, nodeName, tosca.NodeStateError.String())
 		if err != nil {
 			log.Printf("Deployment id: %q, Task id: %q, Failed to set status for node %q: %+v", t.targetID, t.taskID, nodeName, err)
 		}
-		checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusFAILED)
-		return
+		return errors.Wrapf(err, "Command TaskExecution failed for node %q", nodeName)
 	}
 	checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusDONE)
 	metrics.IncrCounter(metricsutil.CleanupMetricKey([]string{"executor", "operation", t.targetID, nodeType, op.Name, "successes"}), 1)
+	return nil
 }
 
 func (w *worker) endAction(ctx context.Context, action *prov.Action, actionErr error) {
@@ -576,6 +569,7 @@ func (w *worker) runDeploy(ctx context.Context, t *taskExecution) {
 		log.Printf("%+v", err)
 		w.checkAndSetDeploymentStatus(ctx, t.targetID, deployments.DEPLOYMENT_FAILED)
 		checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusFAILED)
+		events.PublishAndLogWorkflowStatusChange(ctx, w.consulClient.KV(), t.targetID, t.taskID, "install", tasks.TaskStatusFAILED.String())
 		return
 	}
 	if wfDone {
@@ -585,6 +579,7 @@ func (w *worker) runDeploy(ctx context.Context, t *taskExecution) {
 			log.Printf("%+v", err)
 			return
 		}
+		events.PublishAndLogWorkflowStatusChange(ctx, w.consulClient.KV(), t.targetID, t.taskID, "install", tasks.TaskStatusDONE.String())
 	}
 }
 
@@ -593,15 +588,20 @@ func (w *worker) runUndeploy(ctx context.Context, t *taskExecution) {
 	if err != nil {
 		log.Printf("Deployment id: %q, Task id: %q, Failed to get deployment status: %+v", t.targetID, t.taskID, err)
 		checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusFAILED)
+		events.PublishAndLogWorkflowStatusChange(ctx, w.consulClient.KV(), t.targetID, t.taskID, "uninstall", tasks.TaskStatusFAILED.String())
 		return
 	}
 	if status != deployments.UNDEPLOYED {
-		w.setDeploymentStatus(ctx, t.targetID, deployments.UNDEPLOYMENT_IN_PROGRESS)
+		if status != deployments.UNDEPLOYMENT_IN_PROGRESS {
+			w.setDeploymentStatus(ctx, t.targetID, deployments.UNDEPLOYMENT_IN_PROGRESS)
+		}
+
 		wfDone, err := w.runWorkflowStep(ctx, t, "uninstall", true)
 		if err != nil {
 			log.Printf("%+v", err)
 			w.checkAndSetDeploymentStatus(ctx, t.targetID, deployments.UNDEPLOYMENT_FAILED)
 			checkAndSetTaskStatus(ctx, t.kv, t.taskID, t.step, tasks.TaskStatusFAILED)
+			events.PublishAndLogWorkflowStatusChange(ctx, w.consulClient.KV(), t.targetID, t.taskID, "uninstall", tasks.TaskStatusFAILED.String())
 			return
 		}
 		if wfDone {
@@ -614,6 +614,7 @@ func (w *worker) runUndeploy(ctx context.Context, t *taskExecution) {
 			if t.taskType == tasks.TaskTypePurge {
 				w.runPurge(ctx, t)
 			}
+			events.PublishAndLogWorkflowStatusChange(ctx, w.consulClient.KV(), t.targetID, t.taskID, "uninstall", tasks.TaskStatusDONE.String())
 		}
 	} else if t.taskType == tasks.TaskTypePurge {
 		w.runPurge(ctx, t)