+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+### ENHANCEMENTS
+
+* Generate Alien 2.1-compatible events ([GH-148](https://github.com/ystia/yorc/issues/148))
+
 ### BUG FIXES
 
 * Bootstrap of HA setup fails on GCP, at step configuring the NFS Client component [GH-218](https://github.com/ystia/yorc/issues/218))